Break up axios API functions into their respective modules
