+++ lib/airtime.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Joi = require('@hapi/joi');
-const { initializeAxios } = require('./utils/initializeAxios');
+const initializeAxios = require('./customAxios');
 
 class Airtime {
     constructor(config) {
@@ -24,20 +24,21 @@ class Airtime {
                 amount: `${r.currencyCode} ${r.amount}`,
             }));
 
-            initializeAxios(this.config, false)
-                .sendAirtimeRequest({
-                    recipients: JSON.stringify(recipients),
-                })
-                .then(function (response) {
-                    if (response.status === 201) {
-                        resolve(response.data);
-                    } else {
-                        reject(response.data || response.error);
-                    }
-                })
-                .catch(function (err) {
-                    reject(err);
-                });
+            const customAxios = initializeAxios(this.config, false);
+            
+            customAxios.airtime.sendAirtimeRequest({
+                recipients: JSON.stringify(recipients),
+            })
+            .then(function (response) {
+                if (response.status === 201) {
+                    resolve(response.data);
+                } else {
+                    reject(response.data || response.error);
+                }
+            })
+            .catch(function (err) {
+                reject(err);
+            });
         });        
     }
