+++ lib/application.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { initializeAxios } = require('./utils/initializeAxios');
+const initializeAxios = require('./customAxios');
 
 class Application {
     constructor(config) {
@@ -9,18 +9,19 @@ class Application {
 
     fetchApplicationData() {
         return new Promise((resolve, reject) => {
-            initializeAxios(this.config)
-                .sendGetApplicationDataRequest()
-                .then(function (response) {
-                    if (response.status === 200) {
-                        resolve(response.data);
-                    } else {
-                        reject(response.data || response.error);
-                    }
-                })
-                .catch(function (err) {
-                    reject(err);
-                });
+            const customAxios = initializeAxios(this.config, false);
+            
+            customAxios.application.getApplicationData()
+            .then(function (response) {
+                if (response.status === 200) {
+                    resolve(response.data);
+                } else {
+                    reject(response.data || response.error);
+                }
+            })
+            .catch(function (err) {
+                reject(err);
+            });
 
         });
     }
