+++ lib/customAxios/index.js
@@ -1,8 +1,10 @@
 const axios = require('axios');
-const qs = require('querystring');
+const airtime = require('./airtime');
+const application = require('./application');
 
-const initializeAxios = (config, production = false) => {
+function initializeAxios (config, production = false) {
     const { username, apiKey, format } = config;
+
     const baseURL = production
         ? 'https://api.africastalking.com/version1/'
         : 'https://api.sandbox.africastalking.com/version1';
@@ -14,20 +16,25 @@ const initializeAxios = (config, production = false) => {
                 apiKey,
                 'Content-Type': contentType,
                 Accept: format,
-                data: null,
             },
         });
     };
 
+    const endpoints = {
+        SEND_AIRTIME: '/airtime/send',
+        GET_APPLICATION_DATA: '/user',
+    };
+
+    const opts = {
+        createAxiosInstance,
+        endpoints,
+        username,
+    };
+
     return {
-        sendAirtimeRequest: data => createAxiosInstance().post('/airtime/send', qs.stringify({
-            ...data,
-            username,
-        })),
-        sendGetApplicationDataRequest: () => createAxiosInstance().get('/user', {
-            params: { username },
-        })
+        airtime: airtime(opts),
+        application: application(opts),
     };
 };
 
-exports.initializeAxios = initializeAxios;
+module.exports = initializeAxios;
